% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_cog_mapings.R
\name{format_cog_mapings}
\alias{format_cog_mapings}
\title{Format the COG mappings file for geneplast functions}
\usage{
format_cog_mapings(path)
}
\arguments{
\item{path}{File path to COG mappings (.txt downloaded file)}
}
\value{
A data.frame with three columns: `ssp_id` (species ID from NCBI), `protein_id`
(the ENSEMBL peptide ID) and `cog_id` (COG ID from STRINGdb).
}
\description{
Format the COG mappings file for geneplast functions
}
