% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ssp.R
\name{format_ssp}
\alias{format_ssp}
\title{Retrieve informations about STRINGdb organisms}
\usage{
format_ssp(domain = "all")
}
\arguments{
\item{domain}{Filter the returned data.frame by one of each domain ("bacteria", "eukaryota" and "archaea")
or retrieve all domains (default).}
}
\description{
Retrieve informations about STRINGdb organisms
}
