% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tree.R
\name{generate_tree}
\alias{generate_tree}
\title{Generate a phylotree for geneplast functions}
\usage{
generate_tree(nwk_file, ssp_df)
}
\arguments{
\item{nwk_file}{Newick file containing the phylogenetic tree of interest}

\item{ssp_df}{A data.frame with at least two columns: `ssp_name`, and `ssp_id`.}
}
\value{
A phylotree object ready to be used at geneplast functions
}
\description{
Generate a phylotree for geneplast functions
}
