% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tx_cogs.R
\name{tx_cogs}
\alias{tx_cogs}
\title{Extract orthologous groups from a taxon id}
\usage{
tx_cogs(path, tax_id)
}
\arguments{
\item{path}{Path to COG maps file from STRINGdb.}

\item{tax_id}{A numeric vector of length 1 with the taxon id of interest.}
}
\value{
A data.frame with two columns: the ensembl peptide id and the orthologous groups
}
\description{
Extract orthologous groups from a taxon id
}
\examples{
mouse_cogs <- tx_cogs("~/Downloads/COG.mappings.v10.5.txt", 10090)

}
